refactor: added parametrization for network and Infura key
