+++ .env.example
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_INFURA_API_KEY='<API key>'
+NEXT_PUBLIC_NETWORK=kovan
