+++ .eslintrc.js
@@ -6,6 +6,7 @@ module.exports = {
   env: {
     es2020: true,
     browser: true,
+    node: true,
   },
 
   // Enable ES Module mode.
