+++ components/next-router.js
@@ -25,7 +25,7 @@ export function NextETHLink({ address, children, ...rest }) {
     <Link
       newTab
       href={`https://${
-        router?.query?.network && router?.query?.network !== "mainnet"
+        router?.query?.network && router?.query?.network !== "kovan"
           ? `${router.query.network}.`
           : ""
       }etherscan.io/address/${address}`}
