+++ components/user-settings.stories.js
@@ -45,9 +45,13 @@ const metadata = {
 };
 export default metadata;
 
+const network = process.env.NEXT_PUBLIC_NETWORK || "mainnet";
+
 function Template(args) {
   return (
-    <Web3Provider infuraURL="wss://kovan.infura.io/ws/v3/76223180ca554cad9b16c8879ef4db76">
+    <Web3Provider
+      infuraURL={`wss://${network}.infura.io/ws/v3/${process.env.NEXT_PUBLIC_INFURA_API_KEY}`}
+    >
       <UserSettings {...args} />
     </Web3Provider>
   );
