+++ components/voting-history.stories.js
@@ -6,7 +6,7 @@ import Web3Provider from "./web3-provider";
 
 import KlerosLiquid from "subgraph/abis/kleros-liquid";
 import ProofOfHumanity from "subgraph/abis/proof-of-humanity";
-import { address, klerosLiquidAddress } from "subgraph/config/kovan";
+import { address, klerosLiquidAddress } from "subgraph/config";
 
 const metadata = {
   title: "Arbitration/VotingHistory",
@@ -43,22 +43,24 @@ const metadata = {
 };
 export default metadata;
 
+const network = process.env.NEXT_PUBLIC_NETWORK || "mainnet";
+
 const contracts = [
   {
     name: "proofOfHumanity",
     abi: ProofOfHumanity,
-    address: { kovan: address },
+    address: { [network]: address },
   },
   {
     name: "klerosLiquid",
     abi: KlerosLiquid,
-    address: { kovan: klerosLiquidAddress },
+    address: { [network]: klerosLiquidAddress },
   },
 ];
 function Template(args) {
   return (
     <Web3Provider
-      infuraURL="wss://kovan.infura.io/ws/v3/76223180ca554cad9b16c8879ef4db76"
+      infuraURL={`wss://${network}.infura.io/ws/v3/${process.env.INFURA_API_KEY}`}
       contracts={contracts}
     >
       <ArchonProvider>
