+++ components/web3-provider.stories.mdx
@@ -13,7 +13,7 @@ import { Web3Provider } from "@kleros/components";
 
 import KlerosLiquid from "subgraph/abis/kleros-liquid";
 import ProofOfHumanity from "subgraph/abis/proof-of-humanity";
-import { address, klerosLiquidAddress } from "subgraph/config/kovan";
+import { address, klerosLiquidAddress } from "subgraph/config";
 
 const contracts = [
   {
