+++ netlify.toml
@@ -1,0 +1,3 @@
+[build.environment]
+  NEXT_PUBLIC_INFURA_API_KEY='76223180ca554cad9b16c8879ef4db76'
+  NEXT_PUBLIC_NETWORK='mainnet'
