+++ subgraph/config/index.js
@@ -1,0 +1,10 @@
+import kovan from "./kovan";
+import mainnet from "./mainnet";
+
+const configs = { kovan, mainnet };
+
+export const network = process.env.NEXT_PUBLIC_NETWORK || "mainnet";
+
+export const { address, block, klerosLiquidAddress, UBIAddress } = configs[
+  network
+];
